+++ client/client_test.go
@@ -184,95 +184,6 @@ func TestClientNetworkInfoById(t *testing.T) {
 	}
 }
 
-func TestClientNetworkServiceInvalidCommand(t *testing.T) {
-	var out, errOut bytes.Buffer
-	cli := NewNetworkCli(&out, &errOut, callbackFunc)
-
-	err := cli.Cmd("docker", "network", "service", "invalid")
-	if err == nil {
-		t.Fatalf("Passing invalid commands must fail")
-	}
-}
-
-func TestClientNetworkServiceCreate(t *testing.T) {
-	var out, errOut bytes.Buffer
-	cli := NewNetworkCli(&out, &errOut, callbackFunc)
-
-	err := cli.Cmd("docker", "network", "service", "create", mockServiceName, mockNwName)
-	if err != nil {
-		t.Fatal(err.Error())
-	}
-}
-
-func TestClientNetworkServiceRm(t *testing.T) {
-	var out, errOut bytes.Buffer
-	cli := NewNetworkCli(&out, &errOut, callbackFunc)
-
-	err := cli.Cmd("docker", "network", "service", "rm", mockServiceName, mockNwName)
-	if err != nil {
-		t.Fatal(err.Error())
-	}
-}
-
-func TestClientNetworkServiceLs(t *testing.T) {
-	var out, errOut bytes.Buffer
-	cli := NewNetworkCli(&out, &errOut, callbackFunc)
-
-	err := cli.Cmd("docker", "network", "service", "ls", mockNwName)
-	if err != nil {
-		t.Fatal(err.Error())
-	}
-	if out.String() != string(mockServiceListJSON) {
-		t.Fatal("Network service ls command fail to return the expected list")
-	}
-}
-
-func TestClientNetworkServiceInfo(t *testing.T) {
-	var out, errOut bytes.Buffer
-	cli := NewNetworkCli(&out, &errOut, callbackFunc)
-
-	err := cli.Cmd("docker", "network", "service", "info", mockServiceName, mockNwName)
-	if err != nil {
-		t.Fatal(err.Error())
-	}
-	if out.String() != string(mockServiceJSON) {
-		t.Fatal("Network info command fail to return the expected object")
-	}
-}
-
-func TestClientNetworkServiceInfoById(t *testing.T) {
-	var out, errOut bytes.Buffer
-	cli := NewNetworkCli(&out, &errOut, callbackFunc)
-
-	err := cli.Cmd("docker", "network", "service", "info", mockServiceID, mockNwID)
-	if err != nil {
-		t.Fatal(err.Error())
-	}
-	if out.String() != string(mockServiceJSON) {
-		t.Fatal("Network info command fail to return the expected object")
-	}
-}
-
-func TestClientNetworkServiceJoin(t *testing.T) {
-	var out, errOut bytes.Buffer
-	cli := NewNetworkCli(&out, &errOut, callbackFunc)
-
-	err := cli.Cmd("docker", "network", "service", "join", mockContainerID, mockServiceName, mockNwName)
-	if err != nil {
-		t.Fatal(err.Error())
-	}
-}
-
-func TestClientNetworkServiceLeave(t *testing.T) {
-	var out, errOut bytes.Buffer
-	cli := NewNetworkCli(&out, &errOut, callbackFunc)
-
-	err := cli.Cmd("docker", "network", "service", "leave", mockContainerID, mockServiceName, mockNwName)
-	if err != nil {
-		t.Fatal(err.Error())
-	}
-}
-
 // Docker Flag processing in flag.go uses os.Exit() frequently, even for --help
 // TODO : Handle the --help test-case in the IT when CLI is available
 /*
